+++ includes/admin/class-import-page.php
@@ -7,6 +7,10 @@
 
 namespace CourtneyRDev\SocialWebmentionImporter\Admin;
 
+if ( ! defined( 'ABSPATH' ) ) {
+	exit;
+}
+
 use CourtneyRDev\SocialWebmentionImporter\Import\Comment_Importer;
 use CourtneyRDev\SocialWebmentionImporter\Import\Preview_Record;
 use CourtneyRDev\SocialWebmentionImporter\Plugin;